Add enums to make it more clear

+++ src/main.cpp
@@ -20,6 +20,7 @@ int main(int argc, char* argv[])
     // add unit tests
     // add representative set of matrices
     // use enum in mtx reader?
+    // convert FILE to streams
 
     if (argc != 2) 
     {
@@ -35,9 +36,9 @@ int main(int argc, char* argv[])
         exit(1);
     }
 
-    COO<int> coo = readCoo<int>(f);
+    COO<double> coo = readCoo<double>(f);
     coo.sort();
-    CSR<int> csr = convertCOOToCSR(coo);
+    CSR<double> csr = convertCOOToCSR(coo);
 
     std::cout << "mtx rows: " << coo.num_rows << "\n";
     std::cout << "mtx cols: " << coo.num_cols << "\n";
